Denormalize Queues utility class
