+++ rx-java/rxgrpc-stub/src/main/java/com/salesforce/rxgrpc/stub/RxSubscriberAndServerProducer.java
@@ -32,11 +32,11 @@ public class RxSubscriberAndServerProducer<T>
 
             if (m == QueueSubscription.SYNC) {
                 sourceMode = QueueSubscription.SYNC;
-                queue = Queues.toQueue(f);
+                queue = new SimpleQueueAdapter<T>(f);
                 done = true;
             } else if (m == QueueSubscription.ASYNC) {
                 sourceMode = QueueSubscription.ASYNC;
-                queue = Queues.toQueue(f);
+                queue = new SimpleQueueAdapter<T>(f);
             }
         }
     }
